Home and fun facts html and css files
Questions:
How to make it responsive?
How to link within a page (e.g. “About” in navbar to link to the about section)?
How to align the cards closer together and center them?

Comments:
Sorry had to take the twirl animation on the social media buttons out  I made them static for now but we can play with that if we have time.
All the links are functional yay! (so they link to other pager, e.g. from home to fun facts).
Used bootstrap to make the card flips yay!! 

To do:
Make it responsive (open to ideas as for how to do that).
Add the content (a bit about us).

Let me know if the page loads and looks ok on your laptops!